requirements.txt
altair==5.5.0
anyio==4.9.0
apsw==3.49.1.0
apswutils==0.0.2
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
curl_cffi==0.11.1
et_xmlfile==2.0.0
fa6-icons==0.0.2
fastcore==1.8.1
fastlite==0.1.3
fh-altair==0.1.0
frozendict==2.4.6
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
multitasking==0.0.11
narwhals==1.37.1
numpy==2.2.5
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
peewee==3.18.1
platformdirs==4.3.8
protobuf==6.31.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-fasthtml==0.12.14
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
yfinance==0.2.61
